+++ derive/src/trait_bounds.rs
@@ -18,10 +18,14 @@ use syn::{
     parse_quote,
     punctuated::Punctuated,
     spanned::Spanned,
-    visit::Visit,
+    visit::{
+        self,
+        Visit,
+    },
     Generics,
     Result,
     Type,
+    TypePath,
     WhereClause,
 };
 
@@ -32,6 +36,8 @@ pub fn make_where_clause<'a>(
     input_ident: &'a Ident,
     generics: &'a Generics,
     data: &'a syn::Data,
+    scale_info: &Ident,
+    parity_scale_codec: &Ident,
 ) -> Result<WhereClause> {
     let mut where_clause = generics.where_clause.clone().unwrap_or_else(|| {
         WhereClause {
@@ -39,6 +45,11 @@ pub fn make_where_clause<'a>(
             predicates: Punctuated::new(),
         }
     });
+    for lifetime in generics.lifetimes() {
+        where_clause
+            .predicates
+            .push(parse_quote!(#lifetime: 'static))
+    }
 
     let type_params = generics.type_params();
     let ty_params_ids = type_params
@@ -51,16 +62,27 @@ pub fn make_where_clause<'a>(
 
     let types = collect_types_to_bind(input_ident, data, &ty_params_ids)?;
 
-    types.into_iter().for_each(|ty| {
-        where_clause
-            .predicates
-            .push(parse_quote!(#ty : ::scale_info::TypeInfo + 'static))
+    types.into_iter().for_each(|(ty, is_compact)| {
+        // Compact types need extra bounds, T: HasCompact and <T as
+        // HasCompact>::Type: TypeInfo + 'static
+        if is_compact {
+            where_clause
+                .predicates
+                .push(parse_quote!(#ty : :: #parity_scale_codec ::HasCompact));
+            where_clause
+                .predicates
+                .push(parse_quote!(<#ty as :: #parity_scale_codec ::HasCompact>::Type : :: #scale_info ::TypeInfo + 'static));
+        } else {
+            where_clause
+                .predicates
+                .push(parse_quote!(#ty : :: #scale_info ::TypeInfo + 'static));
+        }
     });
 
     generics.type_params().into_iter().for_each(|type_param| {
         let ident = type_param.ident.clone();
         let mut bounds = type_param.bounds.clone();
-        bounds.push(parse_quote!(::scale_info::TypeInfo));
+        bounds.push(parse_quote!(:: #scale_info ::TypeInfo));
         bounds.push(parse_quote!('static));
         where_clause
             .predicates
@@ -94,14 +116,44 @@ fn type_contains_idents(ty: &Type, idents: &[Ident]) -> bool {
     visitor.result
 }
 
-/// Returns all types that must be added to the where clause with the respective
-/// trait bound.
+/// Checks if the given type or any containing type path starts with the given ident.
+fn type_or_sub_type_path_starts_with_ident(ty: &Type, ident: &Ident) -> bool {
+    // Visits the ast and checks if the a type path starts with the given ident.
+    struct TypePathStartsWithIdent<'a> {
+        result: bool,
+        ident: &'a Ident,
+    }
+
+    impl<'a, 'ast> Visit<'ast> for TypePathStartsWithIdent<'a> {
+        fn visit_type_path(&mut self, i: &'ast TypePath) {
+            if i.qself.is_none() {
+                if let Some(segment) = i.path.segments.first() {
+                    if &segment.ident == self.ident {
+                        self.result = true;
+                        return
+                    }
+                }
+            }
+            visit::visit_type_path(self, i);
+        }
+    }
+
+    let mut visitor = TypePathStartsWithIdent {
+        result: false,
+        ident,
+    };
+    visitor.visit_type(ty);
+    visitor.result
+}
+
+/// Returns all types that must be added to the where clause with a boolean
+/// indicating if the field is [`scale::Compact`] or not.
 fn collect_types_to_bind(
     input_ident: &Ident,
     data: &syn::Data,
     ty_params: &[Ident],
-) -> Result<Vec<Type>> {
-    let types_from_fields = |fields: &Punctuated<syn::Field, _>| -> Vec<syn::Type> {
+) -> Result<Vec<(Type, bool)>> {
+    let types_from_fields = |fields: &Punctuated<syn::Field, _>| -> Vec<(Type, bool)> {
         fields
             .iter()
             .filter(|field| {
@@ -110,9 +162,9 @@ fn collect_types_to_bind(
                 &&
                 // Remove all remaining types that start/contain the input ident
                 // to not have them in the where clause.
-                !type_contains_idents(&field.ty, &[input_ident.clone()])
+                !type_or_sub_type_path_starts_with_ident(&field.ty, &input_ident)
             })
-            .map(|f| f.ty.clone())
+            .map(|f| (f.ty.clone(), super::is_compact(f)))
             .collect()
     };