+++ test_suite/tests/derive.rs
@@ -11,16 +11,18 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
+
 #![cfg_attr(not(feature = "std"), no_std)]
 
-use scale_info::prelude::{
-    boxed::Box,
-    marker::PhantomData,
-};
-
 use pretty_assertions::assert_eq;
+use scale::Encode;
 use scale_info::{
     build::*,
+    prelude::{
+        boxed::Box,
+        marker::PhantomData,
+        vec::Vec,
+    },
     tuple_meta_type,
     Path,
     Type,
@@ -234,6 +236,111 @@ fn associated_types_derive_without_bounds() {
 }
 
 #[test]
+fn associated_types_named_like_the_derived_type_works() {
+    trait Types {
+        type Assoc;
+    }
+    #[allow(unused)]
+    #[derive(TypeInfo)]
+    struct Assoc<T: Types> {
+        a: Vec<T::Assoc>,
+        b: Vec<<T>::Assoc>,
+        c: T::Assoc,
+        d: <T>::Assoc,
+    }
+
+    #[derive(TypeInfo)]
+    enum ConcreteTypes {}
+    impl Types for ConcreteTypes {
+        type Assoc = bool;
+    }
+
+    let struct_type = Type::builder()
+        .path(Path::new("Assoc", "derive"))
+        .type_params(tuple_meta_type!(ConcreteTypes))
+        .composite(
+            Fields::named()
+                .field_of::<Vec<bool>>("a", "Vec<T::Assoc>")
+                .field_of::<Vec<bool>>("b", "Vec<<T>::Assoc>")
+                .field_of::<bool>("c", "T::Assoc")
+                .field_of::<bool>("d", "<T>::Assoc"),
+        );
+
+    assert_type!(Assoc<ConcreteTypes>, struct_type);
+}
+
+#[test]
+fn scale_compact_types_work_in_structs() {
+    #[allow(unused)]
+    #[derive(Encode, TypeInfo)]
+    struct Dense {
+        a: u8,
+        #[codec(compact)]
+        b: u16,
+    }
+
+    let ty_alt = Type::builder()
+        .path(Path::new("Dense", "derive"))
+        .composite(
+            Fields::named()
+                .field_of::<u8>("a", "u8")
+                .compact_of::<u16>("b", "u16"),
+        );
+    assert_type!(Dense, ty_alt);
+}
+
+#[test]
+fn scale_compact_types_work_in_enums() {
+    #[allow(unused)]
+    #[derive(Encode, TypeInfo)]
+    enum MutilatedMultiAddress<AccountId, AccountIndex> {
+        Id(AccountId),
+        Index(#[codec(compact)] AccountIndex),
+        Address32([u8; 32]),
+    }
+
+    let ty = Type::builder()
+        .path(Path::new("MutilatedMultiAddress", "derive"))
+        .type_params(tuple_meta_type!(u8, u16))
+        .variant(
+            Variants::with_fields()
+                .variant("Id", Fields::unnamed().field_of::<u8>("AccountId"))
+                .variant("Index", Fields::unnamed().compact_of::<u16>("AccountIndex"))
+                .variant(
+                    "Address32",
+                    Fields::unnamed().field_of::<[u8; 32]>("[u8; 32]"),
+                ),
+        );
+
+    assert_type!(MutilatedMultiAddress<u8, u16>, ty);
+}
+
+#[test]
+fn type_parameters_with_default_bound_works() {
+    trait Formy {
+        type Tip;
+    }
+    #[derive(TypeInfo)]
+    pub enum MetaFormy {}
+    impl Formy for MetaFormy {
+        type Tip = u8;
+    }
+
+    #[allow(unused)]
+    #[derive(TypeInfo)]
+    struct Bat<TTT: Formy = MetaFormy> {
+        one: TTT,
+    }
+
+    let ty = Type::builder()
+        .path(Path::new("Bat", "derive"))
+        .type_params(tuple_meta_type!(MetaFormy))
+        .composite(Fields::named().field_of::<MetaFormy>("one", "TTT"));
+
+    assert_type!(Bat<MetaFormy>, ty);
+}
+
+#[test]
 fn whitespace_scrubbing_works() {
     #[allow(unused)]
     #[derive(TypeInfo)]
@@ -254,6 +361,9 @@ fn ui_tests() {
     let t = trybuild::TestCases::new();
     t.compile_fail("tests/ui/fail_missing_derive.rs");
     t.compile_fail("tests/ui/fail_unions.rs");
+    t.compile_fail("tests/ui/fail_use_codec_attrs_without_deriving_encode.rs");
+    t.compile_fail("tests/ui/fail_with_invalid_codec_attrs.rs");
+    t.pass("tests/ui/pass_with_valid_codec_attrs.rs");
     t.pass("tests/ui/pass_non_static_lifetime.rs");
     t.pass("tests/ui/pass_self_referential.rs");
     t.pass("tests/ui/pass_basic_generic_type.rs");