+++ test_suite/tests/ui/fail_use_codec_attrs_without_deriving_encode.rs
@@ -1,0 +1,14 @@
+use scale_info::TypeInfo;
+
+#[derive(TypeInfo)]
+struct AttrValidation {
+    a: u8,
+    #[codec(skip)]
+    b: u16,
+}
+
+fn assert_type_info<T: TypeInfo + 'static>() {}
+
+fn main() {
+    assert_type_info::<AttrValidation>();
+}