+++ CHANGELOG.md
@@ -6,6 +6,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.6.0] - 2021-02-05
+### Added
+- Add a TypeDef to handle Compact types [(#53)](https://github.com/paritytech/scale-info/pull/53)
+- Add feature for enabling decoding [(#59)](https://github.com/paritytech/scale-info/pull/59)
+
+### Fixed
+- Derive: use known crate name aliases [(#61)](https://github.com/paritytech/scale-info/pull/61)
+
 ## [0.5.0] - 2021-01-27
 ### Added
 - Add a new TypeDef variant to handle PhantomData - [(#48)](https://github.com/paritytech/scale-info/pull/48)